Convert some .png to .jpg and rename launcher apks
I would like to checkout the files on my server, and
want to have discover images without file extension
- the OUYA figures out on its own if the file contents
are jpg or png.
This way it is easier to automatically build the game data files,
without having to care about the extension

The SkiSafari launcher apk has the wrong package name;
it must have the same cAsIng the original package has.